Fixed bug with file upload - part 2.

+++ src/main/java/com/smanager/storage/FileSystemStorageService.java
@@ -130,6 +130,9 @@ public class FileSystemStorageService implements StorageService {
     }
 
     public void delete(String path) {
+        if (path.equals("")) {
+            return;
+        }
         try {
             Path filename = rootLocation.resolve(path);
             FileSystemUtils.deleteRecursively(filename);
